+++ changelog.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 * Error during the import when required Az.Resource module is not loaded #336
 * Az.DataFactory module is required to import adftools
+* Added function `Set-AdfToolsAuthToken` to enable changing URL to API when target environment is different than default Global Azure #356
 ### Fixed
 * Error when deleting credentials #403